backend/glace/serializers/flavor_serializer.py
from rest_framework import serializers
from glace.models import Flavor


class FlavorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Flavor
        fields = ["id", "name", "price"]
