backend/requirements.txt
django
psycopg2-binary==2.9.7
pytest
pytest-django
pytest-factoryboy
djangorestframework
